pyproject.toml
[project]
name = "hello-azure-deepseek"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "azure-ai-inference>=1.0.0b9",
    "python-dotenv>=1.0.0",
    "prompt-toolkit>=3.0.0",
]
